Sprint One Push 3
I had a lot of trouble actually getting my information to go to the database. Eventually I figured out that my WebController had a GetMap instead of Post which was causing problems. Then for some reason DateConverter was conflicting the WebController so I moved it to the medical package.
Then I had a issue with the name of the entities so I had to rename some of them, and add the @column field to the id. Now everything works correctly and I have finished my work for sprint one.